Tracing: Use theme colors for selection overlay, improve overlay width
Use theme colors for the selection/hover overlay instead of hard-coded
ones, and grow the border outward instead of inward so narrow items
keep a clearly visible outline.

+++ src/libs/tracing/trackpainterbase.h
@@ -155,7 +155,7 @@ protected:
     struct OverlayStroke {
         QRectF rect;
         QRgb color;
-        float lineWidth;
+        qreal lineWidth;
     };
 
     // --- Shared geometry / overlay computation (single-sourced) ---------------
